refactor: Changing the root way we create deepstream clients
Importing and creating the client has changed in order to be less insane when using typescript.

When creating the client you now do:

```typescript
import { DeepstreamClient } from '@deepstream/client'
const dsClient = new DeepstreamClient(url, options)
```

instead of:

```
const deepstream = require('@deepstream/client')
const dsClient = deepstream(url, options)
```
