+++ src/rpc/rpc-handler.ts
@@ -1,4 +1,4 @@
-import { Services } from '../client'
+import { Services } from '../deepstream-client'
 import { Options } from '../client-options'
 import { EVENT, RPC_ACTION, TOPIC, RPCResult, RPCMessage, Message } from '../constants'
 import { RPC, RPCMakeCallback } from '../rpc/rpc'
