+++ src/rpc/rpc-response.ts
@@ -1,4 +1,4 @@
-import { Services } from '../client'
+import { Services } from '../deepstream-client'
 import { Options } from '../client-options'
 import { RPCMessage, TOPIC, RPC_ACTION, RPCResult } from '../constants'
